+++ src/index.html
@@ -13,6 +13,6 @@
 
   <body>
     <div id="mainRoot" style="height: 100%"></div>
-    <script src="./index.js" type="text/javascript"></script>
+    <script src="./index.tsx"></script>
   </body>
 </html>
