+++ src/index.tsx
@@ -6,11 +6,12 @@ import { store, persistor } from "./store/";
 import { PersistGate } from "redux-persist/integration/react";
 
 const mainRoot = document.getElementById("mainRoot");
+
 ReactDOM.render(
-  <Provider store={store}>
-    <PersistGate loading={null} persistor={persistor}>
-      <ViewManager />
-    </PersistGate>
-  </Provider>,
-  mainRoot
+    <Provider store={store}>
+        <PersistGate loading={null} persistor={persistor}>
+            <ViewManager />
+        </PersistGate>
+    </Provider>,
+    mainRoot
 );
